npm shrinkwrap
Best practice is to shrinkwrap applications and use package-lock.json
for libraries. The purpose of shrinkwrapping is to ensure that all
users of the app have the exact same code installed, which helps when
dealing with bug reports. It also resolves ambiguities as to whose
problem it is if there are security issues in dependencies: it's
upstream's, not users'. Without shrinkwrapping some users may have
patched deps installed and others may not, all depending on exactly when
they ran `npm install`. This is unideal for many reasons; shrinkwrapping
fixes this problem.

Note that since package-lock.json will no longer be generated we remove
it from the git ignore list.